- Moved the UI Config Settings to its own package
